# public/swagger_ui/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Gestão de Propostas",
        "version": "v1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/v1/cliente/{id}": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Lista o cliente de acordo com o ID informado",
                "operationId": "35a96a623fa7d9036a0203a1d5e2cd01",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do Usuário",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exemplo consulta cliente por ID"
                    }
                }
            }
        },
        "/api/v1/cliente": {
            "post": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Insere um cliente novo",
                "operationId": "b189dc13fdf1e0097d688a9176a83c50",
                "requestBody": {
                    "description": "Objeto do cadastro",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "document"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Henrique"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "henrique@gmail.com"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "000.000.000-99"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exemplo inserir cliente"
                    }
                }
            }
        },
        "/api/v1/proposta/{id}": {
            "get": {
                "tags": [
                    "Proposta"
                ],
                "summary": "Lista a proposta de acordo com o ID informado",
                "operationId": "f70319379734d4dcb61b1ab06e7936a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Proposta",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exemplo consulta proposta por ID"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Proposta"
                ],
                "summary": "Edita uma proposta",
                "operationId": "e6077fbc7852e53c914d9ff37dca365c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Proposta",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Objeto do cadastro",
                    "required": true,
                    "content": {
                        "application/json-patch+json:": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "produto",
                                    "valor_mensal",
                                    "versao"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "produto": {
                                        "type": "string",
                                        "example": "Produto de Teste 2"
                                    },
                                    "valor_mensal": {
                                        "type": "decimal",
                                        "example": "16.90"
                                    },
                                    "versao": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exemplo sucesso edição proposta"
                    }
                }
            }
        },
        "/api/v1/proposta/": {
            "get": {
                "tags": [
                    "Proposta"
                ],
                "summary": "Lista de propostas com paginação",
                "operationId": "74b1930de449d19c131715a339f07e9c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número da página",
                        "schema": {
                            "type": "integer",
                            "default": "1",
                            "minimum": "1"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número de registros por página",
                        "schema": {
                            "type": "integer",
                            "default": "20",
                            "maximum": "50",
                            "minimum": "1"
                        }
                    },
                    {
                        "name": "titulo",
                        "in": "query",
                        "description": "Título do produto",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Teste"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status do produto",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "DRAFT",
                                "SUBMITTED",
                                "APPROVED",
                                "REJECTED",
                                "CANCELED"
                            ]
                        },
                        "example": "APPROVED"
                    },
                    {
                        "name": "origem",
                        "in": "query",
                        "description": "Origem do produto",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "APP",
                                "SITE",
                                "API"
                            ]
                        },
                        "example": "API"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exemplo consulta cliente por ID"
                    }
                }
            }
        },
        "/api/v1/proposta": {
            "post": {
                "tags": [
                    "Proposta"
                ],
                "summary": "Insere uma proposta nova",
                "operationId": "3603e11ccf8dc2b36fc9f703ddfa938f",
                "requestBody": {
                    "description": "Objeto do cadastro",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "produto",
                                    "valor_mensal",
                                    "origem"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "produto": {
                                        "type": "string",
                                        "example": "Produto de Teste 2"
                                    },
                                    "valor_mensal": {
                                        "type": "decimal",
                                        "example": "16.90"
                                    },
                                    "origem": {
                                        "type": "string",
                                        "enum": [
                                            "APP",
                                            "SITE",
                                            "API"
                                        ],
                                        "example": "SITE"
                                    },
                                    "versao": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exemplo inserir proposta"
                    }
                }
            }
        },
        "/api/v1/proposta/{id}/submit": {
            "post": {
                "tags": [
                    "Proposta"
                ],
                "summary": "Altera status da proposta para submited",
                "operationId": "1316e30c76b55fe3dd3e20874a9b3c10",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Proposta",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exemplo inserir proposta"
                    }
                }
            }
        },
        "/api/v1/proposta/{id}/approve": {
            "post": {
                "tags": [
                    "Proposta"
                ],
                "summary": "Altera status da proposta para approved",
                "operationId": "acaf193968b219cd53feeb62bd0e4eb9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Proposta",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exemplo inserir proposta"
                    }
                }
            }
        },
        "/api/v1/proposta/{id}/reject": {
            "post": {
                "tags": [
                    "Proposta"
                ],
                "summary": "Altera status da proposta para rejected",
                "operationId": "597ebc060f817bfb96cdc4329b101dcc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Proposta",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exemplo inserir proposta"
                    }
                }
            }
        },
        "/api/v1/proposta/{id}/cancel": {
            "post": {
                "tags": [
                    "Proposta"
                ],
                "summary": "Altera status da proposta para canceled",
                "operationId": "fefeafb93ea429ca2563947a38632bed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Proposta",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exemplo inserir proposta"
                    }
                }
            }
        },
        "/api/v1/proposta/{id}/auditoria": {
            "get": {
                "tags": [
                    "Auditoria"
                ],
                "summary": "Lista de auditorias de acordo com o ID da proposta",
                "operationId": "50dccd94960a08513476f15d075782d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da Proposta",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exemplo consulta auditorias por ID da Proposta"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Cliente"
        },
        {
            "name": "Proposta"
        },
        {
            "name": "Auditoria"
        }
    ]
}